+++ packages/icons/CHANGELOG.md
@@ -3,6 +3,100 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.24.0](https://github.com/eruditorgroup/profi-design-system/compare/@eruditorgroup/profi-icons@0.16.0...@eruditorgroup/profi-icons@0.24.0) (2021-07-06)
+
+
+
+# 0.12.0 (2021-07-06)
+
+
+### Features
+
+* **toolkit & ui:** add times connstants & multiple autosuggest api ([#181](https://github.com/eruditorgroup/profi-design-system/issues/181)) ([d55f0bb](https://github.com/eruditorgroup/profi-design-system/commit/d55f0bb486e70dda584068a97655fe526764162d))
+
+
+
+## 0.10.2 (2021-07-05)
+
+
+### Bug Fixes
+
+* **icons:** Fix icons sizes ([#178](https://github.com/eruditorgroup/profi-design-system/issues/178)) ([a50c4a8](https://github.com/eruditorgroup/profi-design-system/commit/a50c4a8aa403fb982c99e76d0e1ec6ce8ac87035))
+
+
+
+# 0.10.0 (2021-07-05)
+
+
+### Features
+
+* **ui:** Autosuggest & Select ([#172](https://github.com/eruditorgroup/profi-design-system/issues/172)) ([7dc8f9f](https://github.com/eruditorgroup/profi-design-system/commit/7dc8f9f1a2aa737591a5dc399278ac2dafd700d8))
+
+
+
+## 0.9.2 (2021-07-05)
+
+
+### Bug Fixes
+
+* Path versions profi-icons & profi-toolkit ([#170](https://github.com/eruditorgroup/profi-design-system/issues/170)) ([aa4c471](https://github.com/eruditorgroup/profi-design-system/commit/aa4c471e9f84a262769315b2b6049a47ed917e9a))
+
+
+### Features
+
+* **icons:** add Diagram, Messages, Parameters, People, Support icons ([#174](https://github.com/eruditorgroup/profi-design-system/issues/174)) ([477c062](https://github.com/eruditorgroup/profi-design-system/commit/477c062e389f678c5cd50c2ada0995dde20cdfb3))
+
+
+
+
+
+# [0.23.0](https://github.com/eruditorgroup/profi-design-system/compare/@eruditorgroup/profi-icons@0.16.0...@eruditorgroup/profi-icons@0.23.0) (2021-07-06)
+
+
+
+# 0.12.0 (2021-07-06)
+
+
+### Features
+
+* **toolkit & ui:** add times connstants & multiple autosuggest api ([#181](https://github.com/eruditorgroup/profi-design-system/issues/181)) ([d55f0bb](https://github.com/eruditorgroup/profi-design-system/commit/d55f0bb486e70dda584068a97655fe526764162d))
+
+
+
+## 0.10.2 (2021-07-05)
+
+
+### Bug Fixes
+
+* **icons:** Fix icons sizes ([#178](https://github.com/eruditorgroup/profi-design-system/issues/178)) ([a50c4a8](https://github.com/eruditorgroup/profi-design-system/commit/a50c4a8aa403fb982c99e76d0e1ec6ce8ac87035))
+
+
+
+# 0.10.0 (2021-07-05)
+
+
+### Features
+
+* **ui:** Autosuggest & Select ([#172](https://github.com/eruditorgroup/profi-design-system/issues/172)) ([7dc8f9f](https://github.com/eruditorgroup/profi-design-system/commit/7dc8f9f1a2aa737591a5dc399278ac2dafd700d8))
+
+
+
+## 0.9.2 (2021-07-05)
+
+
+### Bug Fixes
+
+* Path versions profi-icons & profi-toolkit ([#170](https://github.com/eruditorgroup/profi-design-system/issues/170)) ([aa4c471](https://github.com/eruditorgroup/profi-design-system/commit/aa4c471e9f84a262769315b2b6049a47ed917e9a))
+
+
+### Features
+
+* **icons:** add Diagram, Messages, Parameters, People, Support icons ([#174](https://github.com/eruditorgroup/profi-design-system/issues/174)) ([477c062](https://github.com/eruditorgroup/profi-design-system/commit/477c062e389f678c5cd50c2ada0995dde20cdfb3))
+
+
+
+
+
 # [0.22.0](https://github.com/eruditorgroup/profi-design-system/compare/@eruditorgroup/profi-icons@0.16.0...@eruditorgroup/profi-icons@0.22.0) (2021-07-06)